用户协议模块 & 日志记录优化 & 文章添加修改 descriptions 写成 description 的bug 修复 & 代码格式化

+++ database/migrations/2019_03_06_150428_create_model_has_tags_table.php
@@ -14,7 +14,7 @@ class CreateModelHasTagsTable extends Migration
     public function up()
     {
         Schema::create('model_has_tags', function (Blueprint $table) {
-            $table->unsignedInteger('tag_id');
+            $table->bigIncrements('tag_id');
             $table->morphs('model');
 
             $table->foreign('tag_id')
